Turn off ctrl-c and ctrl-z signals
ctrl-c sends SIGINT signal to current processor
which causes it to terminate. ctrl-z sends SIGTSTP
to current processor which causes it to suspend.

Here we turn off both signals by bitwise anding
the negation of ISIG to stop the terminal from
understanding these signals.

ISIG is not an input flag