Disable ctrl-v and fix ctrl-m
On some systems, when ctrl-v is typed the terminal
waits for you to type another character and then
sends that character. This is turned off using
the IEXTEN flag which is a local flag.
This also fixes ctrl-o on macOS whose terminal
driver is otherwise set to discard that control
character.

ctrl-m points to the carriage return key but it
is translated to 10 which is the return character.
We turn off this feature using the ICRNL flag which
comes from the input flag. CR stands for carriage
return and NL for newline.