Turn off all output processing
Terminal does similar translation on output side.
Each newline we print is turned into "\r\n".
The terminal requires these characters in order to
move the cursor back to the beginning of the
current line and move the cursor down a line.

Output processing features are turned off using
OPOST flag.

We add \r to line 44 so the cursor moves to the
start of the next line.